emotion_engine/core/emotion_state.py
"""
Emotion State Core Module
Maintains VAD vector, active emotions, momentum, stability
"""
import numpy as np
import yaml
from .emotion_memory import EmotionMemory

class EmotionState:
    def __init__(self, config):
        self.vad = np.array(config.get('vad_base', [0.0, 0.0, 0.0]))
        self.active_emotions = {}  # {emotion: weight}
        self.momentum = np.zeros(3)
        self.stability = 1.0
        self.memory = EmotionMemory()
    def update(self, vad_new, emotions_new):
        # TODO: Implement blending, decay, momentum, stability
        pass
    def save(self, path):
        # TODO: Save state to file
        pass
    @staticmethod
    def load_or_init(path=None, config=None):
        # TODO: Load from file or create new
        return EmotionState(config)
