emotion_engine/requirements.txt
numpy
scikit-learn
matplotlib
seaborn
plotly
pandas
shap
pyyaml
